+++ DodocoTales.Library/StoragedUser/Models/DDCLStoragedBannerLogs.cs
@@ -9,7 +9,7 @@ namespace DodocoTales.Library.StoragedUser.Models
 {
     public class DDCLStoragedBannerLogs
     {
-        public int id { get; set; }
+        public ulong id { get; set; }
         public List<DDCLStoragedRoundLogs> R { get; set; }
 
         // as backup
