+++ DodocoTales.Library/StoragedUser/Models/DDCLStoragedRoundLogs.cs
@@ -8,7 +8,7 @@ namespace DodocoTales.Library.StoragedUser.Models
 {
     public class DDCLStoragedRoundLogs
     {
-        public int epitomizedPathID { get; set; }
+        public ulong epitomizedPathID { get; set; }
         public List<DDCLStoragedGachaLogItem> L { get; set; }
     }
 }
